+++ library/src/main/jni/image/image.c
@@ -135,7 +135,7 @@ bool decode_info(Stream* stream, ImageInfo* info) {
 }
 
 bool decode_buffer(Stream* stream, bool clip, uint32_t x, uint32_t y, uint32_t width,
-    uint32_t height, uint8_t config, uint32_t ratio, BufferContainer* container) {
+    uint32_t height, int32_t config, uint32_t ratio, BufferContainer* container) {
   uint8_t magic[2];
   int8_t format;
